+++ src/com/intellij/idea/plugin/hybris/project/utils/PluginCommon.kt
@@ -28,14 +28,6 @@ object PluginCommon {
     val PLUGIN_ID_JREBEL = Plugin("JRebelPlugin", "https://plugins.jetbrains.com/plugin/4441-jrebel-and-xrebel")
     val PLUGIN_ANT_SUPPORT = Plugin("AntSupport", "https://plugins.jetbrains.com/plugin/23025-ant")
     val PLUGIN_MAVEN = Plugin("org.jetbrains.idea.maven")
-    @JvmStatic
-    val PLUGIN_JAVAEE = Plugin("com.intellij.javaee", "https://plugins.jetbrains.com/plugin/20207-jakarta-ee-platform")
-    @JvmStatic
-    val PLUGIN_JAVAEE_WEB = Plugin("com.intellij.javaee.web", "https://plugins.jetbrains.com/plugin/20216-jakarta-ee-web-servlets")
-    @JvmStatic
-    val PLUGIN_JAVAEE_EL = Plugin("com.intellij.javaee.el", "https://plugins.jetbrains.com/plugin/20208-jakarta-ee-expression-language-el-")
-    @JvmStatic
-    val PLUGIN_SPRING = Plugin("com.intellij.spring", "https://plugins.jetbrains.com/plugin/20221-spring")
     val PLUGIN_KOTLIN = Plugin("org.jetbrains.kotlin", "https://plugins.jetbrains.com/plugin/6954-kotlin")
     val PLUGIN_GROOVY = Plugin("org.intellij.groovy", "https://plugins.jetbrains.com/plugin/1524-groovy")
     val PLUGIN_GRADLE = Plugin("org.jetbrains.plugins.gradle")
@@ -43,7 +35,42 @@ object PluginCommon {
     val PLUGIN_DIAGRAM = Plugin("com.intellij.diagram")
     val PLUGIN_PROPERTIES = Plugin("com.intellij.properties", "https://plugins.jetbrains.com/plugin/11594-properties")
     val PLUGIN_COPYRIGHT = Plugin("com.intellij.copyright", "https://plugins.jetbrains.com/plugin/13114-copyright")
+    val PLUGIN_JAVASCRIPT = Plugin("JavaScript", "https://plugins.jetbrains.com/plugin/22069-javascript-and-typescript")
+    val PLUGIN_INTELLILANG = Plugin("org.intellij.intelliLang", "https://plugins.jetbrains.com/plugin/13374-intellilang")
+
+    @JvmStatic
+    val PLUGIN_JAVAEE = Plugin("com.intellij.javaee", "https://plugins.jetbrains.com/plugin/20207-jakarta-ee-platform")
+
+    @JvmStatic
+    val PLUGIN_JAVAEE_WEB = Plugin("com.intellij.javaee.web", "https://plugins.jetbrains.com/plugin/20216-jakarta-ee-web-servlets")
+
+    @JvmStatic
+    val PLUGIN_JAVAEE_EL = Plugin("com.intellij.javaee.el", "https://plugins.jetbrains.com/plugin/20208-jakarta-ee-expression-language-el-")
+
+    @JvmStatic
+    val PLUGIN_SPRING = Plugin("com.intellij.spring", "https://plugins.jetbrains.com/plugin/20221-spring")
+
     const val SHOW_UNLINKED_GRADLE_POPUP = "show.inlinked.gradle.project.popup"
+
+    val PLUGINS = mapOf(
+        PLUGIN_SPRING.id to PLUGIN_SPRING,
+        PLUGIN_KOTLIN.id to PLUGIN_KOTLIN,
+        PLUGIN_GROOVY.id to PLUGIN_GROOVY,
+        PLUGIN_GRADLE.id to PLUGIN_GRADLE,
+        PLUGIN_DATABASE.id to PLUGIN_DATABASE,
+        PLUGIN_DIAGRAM.id to PLUGIN_DIAGRAM,
+        PLUGIN_PROPERTIES.id to PLUGIN_PROPERTIES,
+        PLUGIN_COPYRIGHT.id to PLUGIN_COPYRIGHT,
+        PLUGIN_JAVAEE_EL.id to PLUGIN_JAVAEE_EL,
+        PLUGIN_JAVAEE_WEB.id to PLUGIN_JAVAEE_WEB,
+        PLUGIN_JAVAEE.id to PLUGIN_JAVAEE,
+        PLUGIN_JAVAEE.id to PLUGIN_JAVAEE,
+        PLUGIN_MAVEN.id to PLUGIN_MAVEN,
+        PLUGIN_ANT_SUPPORT.id to PLUGIN_ANT_SUPPORT,
+        PLUGIN_ID_JREBEL.id to PLUGIN_ID_JREBEL,
+        PLUGIN_JAVASCRIPT.id to PLUGIN_JAVASCRIPT,
+        PLUGIN_INTELLILANG.id to PLUGIN_INTELLILANG,
+    )
 
     @JvmStatic
     fun isPluginActive(plugin: Plugin) = PluginManagerCore.getPlugin(PluginId.getId(plugin.id))